+++ src/resources/location.test.ts
@@ -36,11 +36,43 @@ it('will set user filter using a string', async () => {
   });
 });
 
+it('will set user category filter using a number', async () => {
+  const resource = new Location(mockAxios);
+
+  expect(resource.withinUserCategory(1)).toHaveProperty('filters', {
+    user_category: 1,
+  });
+});
+
+it('will set user category filter using a string', async () => {
+  const resource = new Location(mockAxios);
+
+  expect(resource.withinUserCategory('identifier')).toHaveProperty('filters', {
+    user_category: 'identifier',
+  });
+});
+
 it('will set the invite only filter', async () => {
   const resource = new Location(mockAxios);
 
   expect(resource.invitable()).toHaveProperty('filters', {
     invitable: 1,
+  });
+});
+
+it('will set the invite only resources filter to true by default', async () => {
+  const resource = new Location(mockAxios);
+
+  expect(resource.withInviteOnly()).toHaveProperty('filters', {
+    invite_only_resources: true,
+  });
+});
+
+it('can set the invite only resources filter to false', async () => {
+  const resource = new Location(mockAxios);
+
+  expect(resource.withInviteOnly(false)).toHaveProperty('filters', {
+    invite_only_resources: false,
   });
 });
 
@@ -162,6 +194,8 @@ it('can string all filterable options together', async () => {
       .supporting(MeetingMethods.PHONE_CALL)
       .through('client_view')
       .located({ city: 'Fake City', country: 'FC', region: 'FR' })
+      .withInviteOnly()
+      .withinUserCategory(1)
       .take(5)
       .on(1),
   );
@@ -171,12 +205,14 @@ it('can string all filterable options together', async () => {
     city: 'Fake City',
     country: 'FC',
     invitable: 1,
+    invite_only_resources: true,
     method: MeetingMethods.PHONE_CALL,
     preferred: 1,
     region: 'FR',
     resource: 'client_view',
     services: [1, 2],
     user: 1,
+    user_category: 1,
     virtual: 0,
   });
   expected.toHaveProperty('sortable', 'created');
@@ -206,6 +242,8 @@ it('can get locations with additional parameters', async () => {
     .supporting(MeetingMethods.PHONE_CALL)
     .through('client_view')
     .located({ city: 'Fake City', country: 'FC', region: 'FR' })
+    .withInviteOnly()
+    .withinUserCategory(1)
     .sortBy('created')
     .take(5)
     .on(1)
@@ -219,11 +257,13 @@ it('can get locations with additional parameters', async () => {
       'filter[client_view_meeting_method]': MeetingMethods.PHONE_CALL,
       'filter[country]': 'FC',
       'filter[invite_only]': 1,
+      'filter[invite_only_resources]': 1,
       'filter[preferred]': 1,
       'filter[province]': 'FR',
       'filter[service]': [1, 2],
       'filter[resource]': 'client_view',
       'filter[user]': 1,
+      'filter[user_category]': 1,
       'filter[virtual]': 0,
       limit: 5,
       page: 1,