+++ src/resources/location.ts
@@ -18,11 +18,13 @@ export interface LocationFilter {
   [key: string]: any;
   assigned?: boolean;
   invitable?: number;
+  invite_only_resources?: boolean;
   preferred?: number;
   method?: number;
   services?: number | number[] | string | string[];
   resource?: string;
   user?: number | string;
+  user_category?: number | string;
   virtual?: number;
 }
 
@@ -32,11 +34,13 @@ export interface LocationParameters {
   client_view_meeting_method?: number;
   country?: string;
   invite_only?: number;
+  invite_only_resources?: number;
   preferred?: number;
   province?: string;
   service?: number | number[] | string | string[];
   resource?: string;
   user?: number | string;
+  user_category?: number | string;
   virtual?: number;
 }
 
@@ -64,6 +68,10 @@ export interface LocationResource extends Pageable, ConditionalResource {
   through(resource: string): this;
 
   virtual(): this;
+
+  withInviteOnly(inviteOnlyResources?: boolean): this;
+
+  withinUserCategory(userCategory: number | string): this;
 }
 
 export default class Location extends Conditional implements LocationResource {
@@ -208,6 +216,18 @@ export default class Location extends Conditional implements LocationResource {
     return this;
   }
 
+  public withInviteOnly(inviteOnlyResources: boolean = true): this {
+    this.filters.invite_only_resources = inviteOnlyResources;
+
+    return this;
+  }
+
+  public withinUserCategory(userCategory: number | string): this {
+    this.filters.user_category = userCategory;
+
+    return this;
+  }
+
   protected params(): LocationParameters {
     const params: LocationParameters = {};
 
@@ -225,6 +245,10 @@ export default class Location extends Conditional implements LocationResource {
 
     if (typeof this.filters.invitable !== 'undefined') {
       params.invite_only = this.filters.invitable;
+    }
+
+    if (this.filters.invite_only_resources) {
+      params.invite_only_resources = Number(this.filters.invite_only_resources);
     }
 
     if (typeof this.filters.method !== 'undefined') {
@@ -249,6 +273,10 @@ export default class Location extends Conditional implements LocationResource {
 
     if (typeof this.filters.user !== 'undefined') {
       params.user = this.filters.user;
+    }
+
+    if (typeof this.filters.user_category !== 'undefined') {
+      params.user_category = this.filters.user_category;
     }
 
     if (typeof this.filters.virtual !== 'undefined') {