+++ src/resources/service.ts
@@ -17,11 +17,14 @@ export interface ServiceFilter {
   category?: number | string;
   group?: number;
   invitable?: number;
+  invite_only_resources?: boolean,
   location?: number | string;
+  location_category?: number | string;
   method?: number;
   preferred?: number;
   resource?: string;
   user?: number | string;
+  user_category?: number | string;
 }
 
 export interface ServiceParameters {
@@ -30,11 +33,14 @@ export interface ServiceParameters {
   client_view_meeting_method?: number;
   group?: number;
   invite_only?: number;
+  invite_only_resources?: number,
   location?: number | string;
+  location_category?: number | string;
   preferred?: number;
   province?: string;
   resource?: string;
   user?: number | string;
+  user_category?: number | string;
 }
 
 export interface ServiceResource extends Pageable, ConditionalResource {
@@ -59,6 +65,12 @@ export interface ServiceResource extends Pageable, ConditionalResource {
   supporting(method: number): this;
 
   through(resource: string): this;
+
+  withInviteOnly(inviteOnlyResources?: boolean): this;
+
+  withinLocationCategory(locationCategory: number | string): this;
+
+  withinUserCategory(userCategory: number | string): this;
 }
 
 export default class Service extends Conditional implements ServiceResource {
@@ -189,6 +201,24 @@ export default class Service extends Conditional implements ServiceResource {
     return this;
   }
 
+  public withInviteOnly(inviteOnlyResources: boolean = true): this {
+    this.filters.invite_only_resources = inviteOnlyResources;
+
+    return this;
+  }
+
+  public withinLocationCategory(locationCategory: number | string): this {
+    this.filters.location_category = locationCategory;
+
+    return this;
+  }
+
+  public withinUserCategory(userCategory: number | string): this {
+    this.filters.user_category = userCategory;
+
+    return this;
+  }
+
   protected params(): ServiceParameters {
     const params: ServiceParameters = {};
 
@@ -208,6 +238,10 @@ export default class Service extends Conditional implements ServiceResource {
       params.invite_only = this.filters.invitable;
     }
 
+    if (this.filters.invite_only_resources) {
+      params.invite_only_resources = Number(this.filters.invite_only_resources);
+    }
+
     if (typeof this.filters.location !== 'undefined') {
       params.location = this.filters.location;
     }
@@ -223,13 +257,21 @@ export default class Service extends Conditional implements ServiceResource {
     if (typeof this.filters.region !== 'undefined') {
       params.province = this.filters.region;
     }
-    
+
     if (typeof this.filters.resource !== 'undefined') {
       params.resource = this.filters.resource;
     }
 
     if (typeof this.filters.user !== 'undefined') {
       params.user = this.filters.user;
+    }
+
+    if (typeof this.filters.location_category !== 'undefined') {
+      params.location_category = this.filters.location_category;
+    }
+
+    if (typeof this.filters.user_category !== 'undefined') {
+      params.user_category = this.filters.user_category;
     }
 
     return params;